# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ISP Ops Platform API
  version: 0.1.0
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
paths:
...
  /sites:
    get:
      summary: List sites
      responses:
        "200":
          description: Site list
    post:
      summary: Create site
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, name, siteType]
              properties:
                code:
                  type: string
                name:
                  type: string
                siteType:
                  type: string
                status:
                  type: string
                reason:
                  type: string
      responses:
        "201":
          description: Site created
        "400":
          description: Invalid payload
        "503":
          description: PostgreSQL required
...
  /sites/map/import:
    post:
      summary: Preview or apply bulk map import
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                preview:
                  type: boolean
                sites:
                  type: array
                links:
                  type: array
                reason:
                  type: string
      responses:
        "200":
          description: Import preview
        "201":
          description: Import applied
        "207":
          description: Import partially applied
        "400":
          description: Invalid payload
  /sites/transport-links:
    post:
      summary: Create or update a site-to-site transport link
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [aSiteCode, zSiteCode]
              properties:
                aSiteCode:
                  type: string
                zSiteCode:
                  type: string
                providerCode:
                  type: string
                circuitCode:
                  type: string
                linkType:
                  type: string
                status:
                  type: string
                capacityMbps:
                  type: integer
                label:
                  type: string
                reason:
                  type: string
      responses:
        "201":
          description: Transport link created or updated
        "400":
          description: Invalid payload
        "404":
          description: Sites not found or PostgreSQL required
  /sites/{code}/downstream-impact:
    get:
      summary: Site downstream impact
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Downstream impacted sites
        "503":
          description: PostgreSQL required
  /sites/{code}/operational-view:
    get:
      summary: Site operational context
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Site inventory, prefixes and circuits
        "404":
          description: Site not found
  /sites/{code}/location:
    patch:
      summary: Update site map coordinates
      security:
        - ApiKeyAuth: []
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [latitude, longitude]
              properties:
                address:
                  type: string
                latitude:
                  type: number
                longitude:
                  type: number
                reason:
                  type: string
      responses:
        "200":
          description: Location updated
        "400":
          description: Invalid payload
        "404":
          description: Site not found or PostgreSQL required
...